backend/requirements.txt
fastapi
firebase-admin
googletrans
requests